analysis/requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
imageio>=2.9.0
torch>=1.7.0,<1.10.0
torchvision>=0.8.0,<0.10.0
scikit-image>=0.17.0
scipy>=1.5.0
opencv-python>=4.5.0
